backend/requirements.txt
Django==5.1
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
djangorestframework-simplejwt
django-rest-auth==0.6.0
dj-rest-auth==6.0.0
django-allauth==0.55.0
cryptography
pillow==10.4.0